Address review comments Signed-off-by: Alekhya Parisha <aparisha@amazon.com>
