--- data-prepper-plugins/saas-source-plugins/office365-source/src/main/java/org/opensearch/dataprepper/plugins/source/office365/configuration/MetadataKeyAttributes.java
@@ -1,5 +1,0 @@
-package org.opensearch.dataprepper.plugins.source.office365.configuration;
-
-public class MetadataKeyAttributes {
-    public static final String CONTENT_TYPE = "contentType";
-}
