+++ settings.gradle
@@ -196,4 +196,4 @@ include 'data-prepper-plugins:saas-source-plugins:jira-source'
 include 'data-prepper-plugins:saas-source-plugins:confluence-source'
 include 'data-prepper-plugins:saas-source-plugins:atlassian-commons'
 include 'data-prepper-plugins:saas-source-plugins:crowdstrike-source'
-include 'data-prepper-plugins:saas-source-plugins:office365-source'
+include 'data-prepper-plugins:saas-source-plugins:microsoft-office365-source'
